pyproject.toml
[tool.poetry]
name = "gpt-2024-2"
version = "0.1.0"
description = ""
authors = ["Nilo Garcia Monteiro <nilo.garmon@gmail.com>","Pedro Igor Gomes de Morais <pedroigorgm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
bitsandbytes = "^0.44.1"
datasets = "2.18.0"
faiss-gpu = "1.7.2"
gradio = "^5.5.0"
huggingface-hub = "^0.26.2"
yake = "^0.4.8"
rake-nltk = "^1.0.6"
joblib = "1.4.2"
numpy = "1.26.4"
pandas = "2.2.1"
pypdf = "4.2.0"
python-dotenv = "1.0.1"
regex = "2024.7.24"
rouge = "1.0.1"
scikit-learn = "1.4.1.post1"
scipy = "1.14.0"
sentence-transformers = "^3.2.1"
torch = "2.2.1"
tqdm = "4.66.2"
transformers = {extras = ["sentencepiece"], version = "^4.46.2"}
typer = "0.12.3"
accelerate = "^0.33.0"
trl = "^0.9.6"
einops = "^0.8.0"
protobuf = "^5.28.2"
gdown = "^5.2.0"
tiktoken = "^0.8.0"
ragas = "^0.2.2"
langchain-huggingface = "^0.1.0"
nltk = "^3.9.1"
nvidia-ml-py3 = "^7.352.0"
rouge-score = "^0.1.2"
evaluate = "^0.4.3"
bert-score = "^0.3.13"
llama-index = "^0.11.22"
lxml = "^5.3.0"
pymupdf = "^1.24.14"
pymupdf4llm = "^0.0.17"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipykernel = "^6.29.5"

[tool.poetry.scripts]
gpt = "gpt_2024_2.cli:_exec_run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false
